FIX: query_params template tag to replace params
This is arguably not a fix... but it brings the tag in line with the
way that I want to use it. Previously the tag would add another entry
for an existing key-value pair instead of replacing the old value.
In the case of the page number, this would lead to a link like
`?page=1&page=2&page=1`. This changes the tag to replace the existing
value instead of building up the array of associated values in the
QueryDict.