Add get_absolute_url that raises an exception to redirect students
I feel like a removing the method entirely would be unnecessarily
confusing to the students. Especially considering that the template
won't throw an AttributeError for the missing method, it will just
substitute for the empty value.

This adds the methods back but they raise an exception that will
direct the students to use the {% url ... %} template tag instead.